+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [1.0.1] - 2019-05-16
+
+- Fix: action for null string.
+- Fix: range definition of symbols. file works with "Follow cursor".
+  
 ## [1.0.0] - 2019-05-10
 
 - Initial release